feat: refactor DatasetAlerts.tsx to call a generic Alerts component
