+++ src/sections/alerts/Alerts.module.scss
@@ -1,0 +1,4 @@
+.container > * {
+  margin-top: 1em;
+  margin-right: 0.5em;
+}
